refactor(core): dedupe TextInput into app::modals
app/mod.rs carried a method-for-method copy of modals::TextInput
(kept alive only by its own tests and the global-search query). Point
GlobalSearchState at the modals version, drop the copy, and move the
edge cases its tests covered (noop edits at buffer edges, clear,
multibyte backspace/delete) into the modals test module.
