+++ src/app/mod.rs
@@ -287,88 +287,6 @@ pub use modals::{AutomationActionKind, AutomationField, TaskField, TriggerKind};
 /// automation prompt, giving the agent CLI time to come up (~3 s).
 const AGENT_BOOT_DELAY_TICKS: u64 = 300;
 
-pub(crate) struct TextInput {
-    pub(crate) buffer: String,
-    pub(crate) cursor: usize,
-}
-
-#[allow(dead_code)] // some editing methods only exercised by tests
-impl TextInput {
-    fn new() -> Self {
-        Self {
-            buffer: String::new(),
-            cursor: 0,
-        }
-    }
-
-    fn insert(&mut self, c: char) {
-        let byte_pos = self.byte_offset();
-        self.buffer.insert(byte_pos, c);
-        self.cursor += 1;
-    }
-
-    fn backspace(&mut self) {
-        if self.cursor > 0 {
-            self.cursor -= 1;
-            let byte_pos = self.byte_offset();
-            self.buffer.remove(byte_pos);
-        }
-    }
-
-    fn delete(&mut self) {
-        let byte_pos = self.byte_offset();
-        if byte_pos < self.buffer.len() {
-            self.buffer.remove(byte_pos);
-        }
-    }
-
-    fn move_left(&mut self) {
-        self.cursor = self.cursor.saturating_sub(1);
-    }
-
-    fn move_right(&mut self) {
-        let char_count = self.buffer.chars().count();
-        if self.cursor < char_count {
-            self.cursor += 1;
-        }
-    }
-
-    fn home(&mut self) {
-        self.cursor = 0;
-    }
-
-    fn end(&mut self) {
-        self.cursor = self.buffer.chars().count();
-    }
-
-    fn clear(&mut self) {
-        self.buffer.clear();
-        self.cursor = 0;
-    }
-
-    fn set(&mut self, value: &str) {
-        self.buffer = value.to_string();
-        self.cursor = value.chars().count();
-    }
-
-    fn value(&self) -> &str {
-        &self.buffer
-    }
-
-    fn cursor_pos(&self) -> usize {
-        self.cursor
-    }
-
-    /// Convert char-based cursor position to byte offset.
-    fn byte_offset(&self) -> usize {
-        self.buffer
-            .char_indices()
-            .nth(self.cursor)
-            .map(|(i, _)| i)
-            .unwrap_or(self.buffer.len())
-    }
-}
-
 pub enum AppMessage {
     KeyPress(KeyCode, KeyModifiers),
     /// Text pasted via the terminal's bracketed paste mode.
@@ -4718,167 +4636,6 @@ mod tests {
 
     // --- TextInput tests ---
 
-    #[test]
-    fn text_input_new_is_empty() {
-        let input = TextInput::new();
-        assert_eq!(input.value(), "");
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
-    #[test]
-    fn text_input_insert_appends_chars() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.insert('c');
-        assert_eq!(input.value(), "abc");
-        assert_eq!(input.cursor_pos(), 3);
-    }
-
-    #[test]
-    fn text_input_insert_at_middle() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('c');
-        input.move_left();
-        input.insert('b');
-        assert_eq!(input.value(), "abc");
-        assert_eq!(input.cursor_pos(), 2);
-    }
-
-    #[test]
-    fn text_input_backspace_removes_before_cursor() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.insert('c');
-        input.backspace();
-        assert_eq!(input.value(), "ab");
-        assert_eq!(input.cursor_pos(), 2);
-    }
-
-    #[test]
-    fn text_input_backspace_at_start_is_noop() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.home();
-        input.backspace();
-        assert_eq!(input.value(), "a");
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
-    #[test]
-    fn text_input_delete_removes_at_cursor() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.insert('c');
-        input.home();
-        input.delete();
-        assert_eq!(input.value(), "bc");
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
-    #[test]
-    fn text_input_delete_at_end_is_noop() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.delete();
-        assert_eq!(input.value(), "a");
-        assert_eq!(input.cursor_pos(), 1);
-    }
-
-    #[test]
-    fn text_input_move_left_and_right() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        assert_eq!(input.cursor_pos(), 2);
-
-        input.move_left();
-        assert_eq!(input.cursor_pos(), 1);
-
-        input.move_right();
-        assert_eq!(input.cursor_pos(), 2);
-    }
-
-    #[test]
-    fn text_input_move_left_at_zero_is_noop() {
-        let mut input = TextInput::new();
-        input.move_left();
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
-    #[test]
-    fn text_input_move_right_at_end_is_noop() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.move_right();
-        assert_eq!(input.cursor_pos(), 1);
-    }
-
-    #[test]
-    fn text_input_home_and_end() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.insert('c');
-
-        input.home();
-        assert_eq!(input.cursor_pos(), 0);
-
-        input.end();
-        assert_eq!(input.cursor_pos(), 3);
-    }
-
-    #[test]
-    fn text_input_clear_resets_buffer_and_cursor() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.clear();
-        assert_eq!(input.value(), "");
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
-    #[test]
-    fn text_input_multibyte_chars() {
-        let mut input = TextInput::new();
-        input.insert('é');
-        input.insert('ñ');
-        assert_eq!(input.value(), "éñ");
-        assert_eq!(input.cursor_pos(), 2);
-
-        input.move_left();
-        input.insert('ü');
-        assert_eq!(input.value(), "éüñ");
-        assert_eq!(input.cursor_pos(), 2);
-    }
-
-    #[test]
-    fn text_input_backspace_multibyte() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('é');
-        input.insert('b');
-        input.backspace();
-        assert_eq!(input.value(), "aé");
-
-        input.backspace();
-        assert_eq!(input.value(), "a");
-    }
-
-    #[test]
-    fn text_input_delete_multibyte() {
-        let mut input = TextInput::new();
-        input.insert('é');
-        input.insert('b');
-        input.home();
-        input.delete();
-        assert_eq!(input.value(), "b");
-        assert_eq!(input.cursor_pos(), 0);
-    }
-
     // --- Session switching tests ---
 
     /// Stub backend that does nothing — for unit tests only.
@@ -5449,25 +5206,6 @@ mod tests {
     }
 
     // --- Role editor tests ---
-
-    #[test]
-    fn text_input_set_replaces_content_and_moves_cursor_to_end() {
-        let mut input = TextInput::new();
-        input.insert('x');
-        input.set("hello");
-        assert_eq!(input.value(), "hello");
-        assert_eq!(input.cursor_pos(), 5);
-    }
-
-    #[test]
-    fn text_input_set_empty_clears() {
-        let mut input = TextInput::new();
-        input.insert('a');
-        input.insert('b');
-        input.set("");
-        assert_eq!(input.value(), "");
-        assert_eq!(input.cursor_pos(), 0);
-    }
 
     #[test]
     fn ctrl_h_cycles_focus_backward_from_terminal() {
